Extract `generateToken` from `login.json.js`
